+++ frontend/lib/app_setting/widget/app_setting_widget.dart
@@ -7,265 +7,10 @@ import 'package:proj/app_setting/app_setting_api.dart';
 part 'app_box.dart';
 part 'app_logo.dart';
 part 'switcher.dart';
-
-Widget title(String text) {
-  return SizedBox(
-    width: double.infinity,
-    child: Text(
-      text,
-      style: AppStyle.header(color: AppStyle.gray[700]!),
-      textAlign: TextAlign.left,
-    ),
-  );
-}
-Widget unitLine(String key, String value, [VoidCallback? onPress]) {
-  return Container(
-    padding: const EdgeInsets.only(top: 8, bottom: 8, right: 8),
-    height: 44,
-    child: Row(children: [
-      SizedBox(
-        width: 96,
-        child: Text(
-          key,
-          style: AppStyle.header(level: 3, color: AppStyle.gray[700]!),
-        ),
-      ),
-      Expanded(
-        child: Text(
-          value,
-          style: AppStyle.body(color: AppStyle.gray[500]!),
-        ),
-      ),
-      if (onPress != null)
-        Container(
-          padding: const EdgeInsets.symmetric(vertical: 4, horizontal: 8),
-          child: TextButton(
-            onPressed: onPress,
-            child: Text(
-              '修改',
-              style: AppStyle.caption(color: AppStyle.teal),
-            ),
-          ),
-        )
-    ]),
-  );
-}
-Widget regardingUS() {
-  return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 16),
-      decoration: BoxDecoration(
-        color: AppStyle.white,
-        borderRadius: BorderRadius.circular(4),
-      ),
-      child: Column(
-        children: [
-          title("關於我們"),
-          const SizedBox(
-            height: 12,
-          ),
-          Divider(
-            height: 1,
-            thickness: 1,
-            color: AppStyle.gray[100],
-          ),
-          const SizedBox(
-            height: 8,
-          ),
-          Container(
-            height: 192,
-            padding: const EdgeInsets.symmetric(vertical: 16),
-            child: Image.asset(
-              "assets/images/logo.png",
-              fit: BoxFit.contain,
-            ),
-          ),
-          const SizedBox(
-            height: 8,
-          ),
-          Text(
-            'fkjweoaonviewn;volasdnklf;dsnfmklsdmnkl;fmsdldkfjkals;djfkldslkf;aklsdfds',
-            style: AppStyle.info(level: 2, color: AppStyle.gray[700]!),
-          ),
-        ]
-      )
-  );       
-}
-
-Widget versionInfo() {
-  return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 16),
-      decoration: BoxDecoration(
-        color: AppStyle.white,
-        borderRadius: BorderRadius.circular(4),
-      ),
-      child: Column(
-        children: [
-          title("版本資訊"),
-          const SizedBox(
-            height: 12,
-          ),
-          Divider(
-            height: 1,
-            thickness: 1,
-            color: AppStyle.gray[100],
-          ),
-          unitLine(
-            "版本",
-            "1.2.8",
-          ),
-        ]
-      )
-  );
-}
-class loginNotification extends StatefulWidget {
-  const loginNotification({super.key});
-
-  @override
-  State<loginNotification> createState() => _loginNotificationState();
-}
-class _loginNotificationState extends State<loginNotification> {
-  bool _isChecked = true;
-  int _responseCode = 400;
-
-  Future<void> _modifyNoti(
-      {String? isLoginMail}) async {
-    try {
-      final int responseCode = await AppSettingAPI.modifyNoti(
-          isLoginMail: isLoginMail);
-      setState(() {
-        _responseCode = responseCode;
-      });
-    } catch (e) {
-      print('API request error: $e');
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 16),
-      decoration: BoxDecoration(
-        color: AppStyle.white,
-        borderRadius: BorderRadius.circular(4),
-      ),
-      child: Column(
-        children: [
-          title("登入通知"),
-          const SizedBox(
-            height: 12,
-          ),
-          Divider(
-            height: 1,
-            thickness: 1,
-            color: AppStyle.gray[100],
-          ),
-          Container(
-            padding: const EdgeInsets.only(top: 8, bottom: 8, right: 8),
-            height: 44,
-            child: Row(children: [
-              Container(
-                width: 146,
-                child: Row(
-                  children: [
-                    SizedBox(
-                      width: 80,
-                      child: Text(
-                        "開啟/關閉",
-                        style: AppStyle.header(level: 3, color: AppStyle.gray[700]!),
-                      ),
-                    ),
-                    const SizedBox(width: 16),
-                    Switcher(
-                      onChanged: (value) async {
-                        setState(() {
-                          _isChecked = value;
-                        });
-                        _isChecked == true ? await _modifyNoti(isLoginMail: "true") :
-                        await _modifyNoti(isLoginMail: "false");
-                        if (_responseCode == 200) {
-                          if (!context.mounted) return;
-                          print("switcher成功");
-                        } else {
-                          if (!context.mounted) return;
-                          print("switcher壞了");
-                        }
-                      },
-                    ),
-                  ],
-                ),
-              ),
-            ])
-          )
-        ]
-      )
-  );
-  }
-}
-
-Widget reportProblem(BuildContext context) {
-  return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 16),
-      decoration: BoxDecoration(
-        color: AppStyle.white,
-        borderRadius: BorderRadius.circular(4),
-      ),
-      child: Column(
-        children: [
-          title("回報問題"),
-          const SizedBox(
-            height: 12,
-          ),
-          Divider(
-            height: 1,
-            thickness: 1,
-            color: AppStyle.gray[100],
-          ),
-          const SizedBox(height: 8),
-          OutlinedButton(
-            onPressed: () {
-              showDialog(
-                context: context,
-                barrierDismissible: false,
-                builder: (context) => const PopProbFB(),
-              );
-            },
-            style: AppStyle.popUpBtn(),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Image.asset("assets/icons/Bell.png"),
-                const SizedBox(
-                  width: 8,
-                ),
-                const Text("問題/反饋")
-              ],
-            ),
-          ),
-        ]
-      )
-  );
-}
-Widget logoutBtn(BuildContext context) {
-  return OutlinedButton(
-    onPressed: () {
-      Navigator.popAndPushNamed(context, '/login');
-    },
-    style: AppStyle.dangerBtn().copyWith(
-      minimumSize: MaterialStateProperty.all<Size>(
-        const Size(95, 40),
-      ),
-    ),
-    child: Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        Image.asset("assets/icons/Sign_out_circle.png"),
-        const SizedBox(
-          width: 8,
-        ),
-        const Text("登出")
-      ],
-    ),
-  );
-}
+part 'title.dart';
+part 'unit_line.dart';
+part 'regarding_us.dart';
+part 'version_info.dart';
+part 'login_notification.dart';
+part 'report_problem.dart';
+part 'logout_btn.dart';
