+++ frontend/lib/app_setting/widget/title.dart
@@ -1,0 +1,12 @@
+part of 'app_setting_widget.dart';
+
+Widget title(String text) {
+  return SizedBox(
+    width: double.infinity,
+    child: Text(
+      text,
+      style: AppStyle.header(color: AppStyle.gray[700]!),
+      textAlign: TextAlign.left,
+    ),
+  );
+}
