+++ frontend/lib/app_setting/widget/version_info.dart
@@ -1,0 +1,28 @@
+part of 'app_setting_widget.dart';
+
+Widget versionInfo() {
+  return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 16),
+      decoration: BoxDecoration(
+        color: AppStyle.white,
+        borderRadius: BorderRadius.circular(4),
+      ),
+      child: Column(
+        children: [
+          title("版本資訊"),
+          const SizedBox(
+            height: 12,
+          ),
+          Divider(
+            height: 1,
+            thickness: 1,
+            color: AppStyle.gray[100],
+          ),
+          unitLine(
+            "版本",
+            "1.2.8",
+          ),
+        ]
+      )
+  );
+}
